# internal/constants/dto/jobPost.go
package dto

import (
	"time"

	"github.com/google/uuid"
	"github.com/pgvector/pgvector-go"
)

type CreateJobPostRequest struct {
	ID          string    `json:"id"`
	UserID      uuid.UUID `json:"user_id"`
	Title       string    `json:"title" binding:"required"`
	Description string    `json:"description" binding:"required"`
	Deadline    time.Time `json:"deadline" binding:"required"`

	Responsibilities []string `json:"responsibilities" binding:"required,dive,required"`
	Requirements     []string `json:"requirements" binding:"required,dive,required"`

	DescriptionEmbedding      pgvector.Vector `json:"description_embedding" swaggerignore:"true"`
	RequirementsEmbedding     pgvector.Vector `json:"requirements_embedding" swaggerignore:"true"`
	ResponsibilitiesEmbedding pgvector.Vector `json:"responsibilities_embedding" swaggerignore:"true"`
}

type CreateJobPostResponse struct {
	ID               string    `json:"id"`
	Title            string    `json:"title"`
	Description      string    `json:"description"`
	Responsibilities []string  `json:"responsibilities"`
	Requirements     []string  `json:"requirements"`
	UserID           uuid.UUID `json:"user_id"`
	Deadline         time.Time `json:"deadline"`
	// CreatedAt   time.Time `json:"created_at"`
}

type GetAllJobPostsResponse struct {
	ID               string    `json:"id"`
	Title            string    `json:"title"`
	Description      string    `json:"description"`
	Responsibilities []string  `json:"responsibilities"`
	Requirements     []string  `json:"requirements"`
	UserID           uuid.UUID `json:"user_id"`
	Deadline         time.Time `json:"deadline"`
}

// UpdateJobPostRequest supports partial updates; omitted fields are ignored.
type UpdateJobPostRequest struct {
	ID               string     `json:"id"`
	Title            *string    `json:"title,omitempty" binding:"omitempty,min=1,max=255"`
	Description      *string    `json:"description,omitempty" binding:"omitempty,min=1"`
	Deadline         *time.Time `json:"deadline,omitempty" binding:"omitempty"`
	Responsibilities *[]string  `json:"responsibilities,omitempty" binding:"omitempty,min=1,dive,required"`
	Requirements     *[]string  `json:"requirements,omitempty" binding:"omitempty,min=1,dive,required"`

	DescriptionEmbedding      *pgvector.Vector `json:"-"`
	RequirementsEmbedding     *pgvector.Vector `json:"-"`
	ResponsibilitiesEmbedding *pgvector.Vector `json:"-"`
}

type UpdateJobPostResponse struct {
	ID               string    `json:"id"`
	Title            string    `json:"title"`
	Description      string    `json:"description"`
	Responsibilities []string  `json:"responsibilities"`
	Requirements     []string  `json:"requirements"`
	UserID           uuid.UUID `json:"user_id"`
	Deadline         time.Time `json:"deadline"`
}